Changed GLUE2GroupID to GLUE2GroupName
Also changed the way the DN is built. This change is backward compatible
as it uses attribute aliases. OpenLDAP automatically renames
GLUE2GroupID to GLUE2GroupName.

The motivation for this change is that GLUE2GroupID is NOT an ID like
GLUE2 describes it, and is not unique (same group can be repeated in
different branches of the tree)

Schema Version updated.
ChangeLog updated.