Changed after LDAP DRAFT review carried on October 2013
This commit syncronizes the schema files with the
LDAP realisation document contents and with some choices in
the XML rendering.

Highlights are: Boolean SYNTAX type discarded in favour
of all DirectoryString types, all objects that are not
ABSTRACT in GFD.147 are turned into STRUCTURAL.

Some objects were incorrectly created as STRUCTURAL
and this has been changed.